+++ src/main/java/io/xpring/common/XRPLNetwork.java
@@ -1,8 +1,13 @@
 package io.xpring.common;
 
+import io.xpring.common.idiomatic.XrplNetwork;
+
 /**
  * Possible networks to resolve.
+ *
+ * @deprecated Please use the idiomatically named {@link XrplNetwork} enum instead.
  */
+@Deprecated
 @SuppressWarnings("checkstyle:AbbreviationAsWordInName")
 public enum XRPLNetwork {
   DEV("devnet"),