+++ src/main/java/io/xpring/xrpl/idiomatic/XrpClientInterface.java
@@ -1,36 +1,29 @@
-package io.xpring.xrpl;
+package io.xpring.xrpl.idiomatic;
 
-import io.xpring.common.XRPLNetwork;
-import io.xpring.xrpl.idiomatic.XrpClient;
-import io.xpring.xrpl.idiomatic.XrpClientInterface;
-import io.xpring.xrpl.model.XRPTransaction;
+import io.xpring.common.idiomatic.XrplNetwork;
+import io.xpring.xrpl.TransactionStatus;
+import io.xpring.xrpl.Wallet;
+import io.xpring.xrpl.model.idiomatic.XrpTransaction;
 
 import java.math.BigInteger;
 import java.util.List;
 
-/**
- * An interface for an {@link XrpClient}.
- *
- * @deprecated Please use the idiomatically named {@link XrpClientInterface} class instead.
- */
-@SuppressWarnings("checkstyle:AbbreviationAsWordInName")
-@Deprecated
-public interface XRPClientInterface {
+public interface XrpClientInterface {
   /**
-   * Retrieve the network that this XRPClient connects to.
+   * Retrieve the network that this XrpClient connects to.
    *
-   * @return The {@link XRPLNetwork} of an {@link XRPClientInterface}
+   * @return The {@link XrplNetwork} of an {@link XrpClientInterface}
    */
-  public XRPLNetwork getNetwork();
+  public XrplNetwork getNetwork();
 
   /**
    * Get the balance of the specified account on the XRP Ledger.
    *
    * @param xrplAccountAddress The X-Address to retrieve the balance for.
    * @return A {@link BigInteger} with the number of drops in this account.
-   * @throws XRPException If the given inputs were invalid.
+   * @throws XrpException If the given inputs were invalid.
    */
-  BigInteger getBalance(final String xrplAccountAddress) throws XRPException;
+  BigInteger getBalance(final String xrplAccountAddress) throws XrpException;
 
   /**
    * Retrieve the transaction status for a Payment given transaction hash.
@@ -41,9 +34,9 @@ public interface XRPClientInterface {
    * </p>
    * @param transactionHash The hash of the transaction.
    * @return The status of the given transaction.
-   * @throws XRPException If the given inputs were invalid.
+   * @throws XrpException If the given inputs were invalid.
    */
-  TransactionStatus getPaymentStatus(String transactionHash) throws XRPException;
+  TransactionStatus getPaymentStatus(String transactionHash) throws XrpException;
 
   /**
    * Transact XRP between two accounts on the ledger.
@@ -52,9 +45,9 @@ public interface XRPClientInterface {
    * @param destinationAddress The X-Address to send the XRP to.
    * @param sourceWallet       The {@link Wallet} which holds the XRP.
    * @return A transaction hash for the payment.
-   * @throws XRPException If the given inputs were invalid.
+   * @throws XrpException If the given inputs were invalid.
    */
-  public String send(BigInteger amount, String destinationAddress, Wallet sourceWallet) throws XRPException;
+  public String send(BigInteger amount, String destinationAddress, Wallet sourceWallet) throws XrpException;
 
   /**
    * Return the history of payments for the given account.
@@ -65,9 +58,9 @@ public interface XRPClientInterface {
    * </p>
    * @param address The address (account) for which to retrieve payment history.
    * @return An array of transactions associated with the account.
-   * @throws XRPException If there was a problem communicating with the XRP Ledger.
+   * @throws XrpException If there was a problem communicating with the XRP Ledger.
    */
-  public List<XRPTransaction> paymentHistory(String address) throws XRPException;
+  public List<XrpTransaction> paymentHistory(String address) throws XrpException;
 
   /**
    * Retrieve the payment transaction corresponding to the given transaction hash.
@@ -76,8 +69,8 @@ public interface XRPClientInterface {
    *       See the `validated` field to make this distinction.
    * </p>
    * @param transactionHash The hash of the transaction to retrieve.
-   * @return An XRPTransaction object representing an XRP Ledger transaction.
+   * @return An XrpTransaction object representing an XRP Ledger transaction.
    * @throws io.grpc.StatusRuntimeException If the transaction hash was invalid.
    */
-  public XRPTransaction getPayment(String transactionHash) throws XRPException;
+  public XrpTransaction getPayment(String transactionHash) throws XrpException;
 }